Add IGSN listing to tinymce plugin
